AAP: Define FLT_ROUNDS as 1 (to nearest)
FLT_ROUNDS by default expands to a call to __builtin_flt_rounds which is
turned into the llvm.flt.rounds intrinsics. This intrinsic attempts to
bitcast the returned i32 but is unable as we do not natively support i32
types.

We do not (yet) support floating point rounding modes so for now this is
disabled.