[CodeGen] Fix __builtin_flt_rounds return value
In LLVM, this has an i32 return value, but in Clang the builtin is
marked as returning an integer. If we rely on the default builtin
handling code on a target with i16 integers, we trigger an assertion
when it tries to bitcast from the LLVM type to the Clang type.