[CodeGen] Use constant int arguments for builtins that require them
In order to emit a function call to the builtins 'returnaddress' and
'frameaddress' where the target has an integer size that is not 32
bits, the argument cannot be assigned a type of 'UnsignedIntTy'.
Instead the argument is emitted as a constant 'Int32Ty', which is the
LLVM type that these builtins expect.