AAP: Add ABI info for AAP to generate correct calls
Using the default ABI info class is not sufficient to generate the
correct calls on the AAP target. Specifically varargs could not be
correctly handled as for aggregate types the caller was generating a
'byval' argument where the callee expected to access the argument
through a pointer.

This change fixes the above issue and provides a concrete
implementation of the ABI for the AAP target and a test to ensure the
generated code is correct.